smtp2telegram-docker/requirements.txt
flask
python-telegram-bot
bleach